Allow filename to contain version replacements
This makes it possible to handle assets whos name contains version
specific parts.

Like a version 74.1 that contains an asset with filename
'asset-74.1.tgz` which can now be referenced using
'asset-%major%.%minor%.tgz - the replacement will be done before
fetching the asset but after version matching. So that it will also work
for a version constraint of `^74.0`