Fetch all releasesm not only first page
This will now fetch all releases for a project and no longer only the
first page. This will take a while on first call as it needs to fetch
the data from the release-list over and over again (for PHPStan that's
about 11 pages of each 50 releases).

After caching the resaponse is much faster though!